add/server: add method return object video/image

+++ server/src/main/java/com/duokoala/server/entity/media/Video.java
@@ -18,5 +18,5 @@ public class Video {
     @GeneratedValue(strategy = GenerationType.UUID)
     String videoId;
     String videoUrl;
-    int videoDuration;
+    double videoDuration;
 }
